Feature/crud user

+++ src/main/java/com/vash/entel/model/entity/Customer.java
@@ -14,7 +14,7 @@ public class Customer {
     private Integer id;
 
     @Column(name = "doc_number", nullable = false)
-    private Integer doc_number;
+    private Long doc_number;
 
     @Column(name = "fullname", nullable = false)
     private String fullname;
